Separate interactor modes to their own file.
